+++ src/Atoms/Media/Media.types.ts
@@ -3,4 +3,5 @@ import { Theme } from '../../theme/theme.types';
 export type Props = {
   as?: keyof JSX.IntrinsicElements | React.ComponentType<{ className?: string }>;
   query: string | ((theme: Theme) => string);
+  children?: React.ReactNode;
 };