+++ src/Atoms/TabTitle/TabTitle.types.ts
@@ -2,4 +2,5 @@ export type Props = {
   active?: boolean;
   height?: number;
   variant?: 'normal' | 'large';
+  children?: React.ReactNode;
 };