+++ src/Organisms/StatusModal/StatusModal.tsx
@@ -3,7 +3,14 @@ import { Box, Button, Flexbox, Modal, Spinner, Typography, OldIcon } from '../..
 import { Props } from './StatusModal.types';
 import { isFunction } from '../../common/utils';
 
-const StatusModal: React.FC<Props> = ({ id = '', loading = false, onClose, options = {} }) => {
+const StatusModal: React.FC<Props> = ({
+  id = '',
+  loading = false,
+  onClose,
+  options = {},
+  fixedBottomMobile = false,
+  confirmColor,
+}) => {
   const [showModal, setShowModal] = useState(false);
 
   const { status, title, text, textConfirm, textCancel } = options || {};
@@ -35,6 +42,7 @@ const StatusModal: React.FC<Props> = ({ id = '', loading = false, onClose, optio
       closeOnBackdropClick
       isStatusModal
       fullScreenMobile={false}
+      fixedBottomMobile={fixedBottomMobile}
     >
       <Box pt={7} pb={7} px={2} sm={{ px: 0, pb: 0, pt: 5 }}>
         <Flexbox
@@ -70,7 +78,11 @@ const StatusModal: React.FC<Props> = ({ id = '', loading = false, onClose, optio
               <Flexbox item>
                 <Box pt={2}>
                   {textConfirm && (
-                    <Button size="l" onClick={() => closeModal(true)}>
+                    <Button
+                      size="l"
+                      onClick={() => closeModal(true)}
+                      color={(t) => (confirmColor ? confirmColor(t) : t.color.cta)}
+                    >
                       {textConfirm}
                     </Button>
                   )}
@@ -91,7 +103,12 @@ const StatusModal: React.FC<Props> = ({ id = '', loading = false, onClose, optio
                 <Flexbox item flex="0">
                   <Box pt={2}>
                     {textConfirm && (
-                      <Button size="l" variant="primary" onClick={() => closeModal(true)}>
+                      <Button
+                        size="l"
+                        variant="primary"
+                        onClick={() => closeModal(true)}
+                        color={(t) => (confirmColor ? confirmColor(t) : t.color.cta)}
+                      >
                         {textConfirm}
                       </Button>
                     )}